refactor(nextwiki): 替换 NextwikiConverter 为 NextwikiStructMapper
- 将所有 NextwikiConverter 的调用替换为 NextwikiStructMapper，提高转换一致性
- 注入并使用 NextwikiStructMapper 作为 MapStruct 转换器实例
- 统一文件版本、文件节点及回收站条目等 DTO 与 VO 转换逻辑
- 简化代码结构，增强类型安全与易维护性

fix(agent): 将温度参数改用 BigDecimal 表示

- 将 AgentDebuggerService 中的 REPLAY_TEMPERATURE 类型由 double 改为 BigDecimal
- 统一 DagOrchestrationExecutor 和 SupervisorAgentExecutor 中温度参数类型为 BigDecimal，避免精度误差

fix(userinfo): 优化默认容量设置以提升效率

- 为 OAuth2 客户端注册表和分端会话限制配置设置默认初始容量常量
- 替换硬编码的容量值为对应常量，增强代码可读性和维护性

fix(literule): 移除对 MapStruct 的依赖，改为手动映射

- 从 ydsl-literule-server 模块中移除 MapStruct 依赖，符合 YDIZ-DDD-004 规范
- 由手动映射代替自动映射，避免跨层依赖问题

fix(rule): 添加日志记录无法识别的规则状态值

- 在 RuleStatusEnum 的 fromName 方法中捕获异常时打印 debug 日志
- 提升系统调试能力，方便排查状态解析失败的问题

fix(chat): 将单价 BigDecimal 转为 double 以兼容 API

- 改进 TokenCostCalculator 中估算费用和实际费用计算逻辑
- 在调用 CostEstimate 方法时，将 BigDecimal 类型单价转换成 double，避免序列化问题

fix(workflow): 修正发布状态查询字段名错误

- 将 FlowDefinition 查询中使用的发布状态字段从 isPublish 改为 publishStatus
- 保证流程定义查询条件正确，避免潜在错误数据过滤错误

+++ ydsz-agent/ydsz-agent-server/src/main/java/com/njydsz/agent/server/agent/DagOrchestrationExecutor.java
@@ -1,5 +1,6 @@
 package com.njydsz.agent.server.agent;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.HashMap;
@@ -469,7 +470,7 @@ public class DagOrchestrationExecutor implements AgentExecutor {
         resolveNodeAgentType(node.getAgentType()),
         node.getPrompt(),
         List.of(),
-        properties.getLlm().getTemperature(),
+        BigDecimal.valueOf(properties.getLlm().getTemperature()),
         properties.getLlm().getMaxTokens(),
         getNodeMaxIterations(node),
         null);
