+++ ydsz-agent/ydsz-agent-server/src/main/java/com/njydsz/agent/server/chat/TokenCostCalculator.java
@@ -66,7 +66,9 @@ public class TokenCostCalculator {
     BigDecimal charRatio = properties.getMemory().getTokenCharRatio();
     int estimatedPromptTokens = Math.max(1, (int) Math.ceil(totalChars / charRatio.doubleValue()));
     BigDecimal unitPrice = resolveUnitPrice(request.getModel());
-    return CostEstimate.estimate(estimatedPromptTokens, request.getMaxTokens(), request.getModel(), unitPrice);
+    // CostEstimate 为 API 请求值对象，最终序列化为 JSON number；此处将 BigDecimal 单价转为 double 传入（API 边界转换）
+    return CostEstimate.estimate(
+        estimatedPromptTokens, request.getMaxTokens(), request.getModel(), unitPrice.doubleValue());
   }
 
   /**
@@ -78,7 +80,7 @@ public class TokenCostCalculator {
    */
   public CostEstimate calculateActual(TokenUsage usage, String model) {
     BigDecimal unitPrice = resolveUnitPrice(model);
-    return CostEstimate.actual(usage, model, unitPrice);
+    return CostEstimate.actual(usage, model, unitPrice.doubleValue());
   }
 
   /**
